Made a start on FDI disk image support, with code to parse the track information.
Removed the code that was to be used for the new SDF image format, as FDI support will replace it.

Added stricter check to CFileDisk, to restrict it to files small enough to fit on a disk.

Changed some types from signed to unsigned, and others to use size_t.