Removed SpectrumROM option as it doesn't belong in a SAM emulator.  Users can still switch the normal SAM ROM for an appropriately modified Spectrum ROM if they wish, which is also true of the real machine.
Default sound latency is now 5 frames, which is the number of extra frames to buffer.  This avoids the problem of someone setting it to zero in the hope of zero latency, which just ain't possible!

Turbo and Paused are now global flags rather than options, since their state shouldn't be saved.

Renamed frequency option/variable to freq, to save on a few characters of typing ;-)

Changed C++ namespace to a class with static functions.