Removed the ROM patching used for fast startup, as the new method runs without video generation to accelerate startup.  Hard disk booting will now be available using Edwin's modified ROM as an external file.
RAM is now initialised in row stripes, to simulate the effect observed on a cold booted SAM.  Not-present RAM is now filled with 0xff rather than 0x00, which seems like a better default.

Added support for a ROM file containing the full 32K ROM, which will ignore the second ROM file if found.  The file may also be a ZX82 image with a 140 byte header, to allow Andy Wright's images to be used unchanged.

ROM images are now read using CStream, so they may be compressed with gzip or zip if desired.

Changed C++ namespace to a class with static functions.