Tweaked power-on defaults to be closer to the real SAM - in particular, the CLUT is no longer pre-loaded.
Implemented the power-on ASIC startup time (~49ms), during which it doesn't respond to I/O.  It's likely the delay at the start of the SAM ROM is giving time for the ASIC to become fully operational.

Rather than unpatching the ROMs when keyboard reads are detected, the new fast start counter is simply reset.

Changed C++ namespace to a class with static functions.