Tweaked the about box, and updated the version number shown to 0.90 WIP (5).
Shortened the text for the 'note' on System tab in the options.

Widened list box on the Fn Keys tab in the options, as it seems to need a few extra pixels on XP to avoid a horizontal scrollbar.

Added ellipsis to the menu options that display a dialogue, to indicate that fact.