Native pixel values are now calculated for all non-palettised surface formats, rather than assuming 24-bit is BGR and 32-bit is (A)RGB.  This fixes a wrong-colour problem on some unusual cards (SGI in this case).
The SAM palette generation has been moved to IO.cpp, since it's common to all platforms.

The screen back-buffer no longer depends on the scanlines option, since it needs to always be full height for the GUI.

The current state of the dimmed palette is now updated whenever the palette is re-created, to ensure we don't lose track when multiple dim/undim events are overlapped.

Changed C++ namespace to class with static functions.