Added preliminary localisation functionality, which enumerates the text on menus/windows to locate strings to substitute.
Added stretch-to-fit to the Display options.  This allowed the windowed mode window to be resized to any size, and will use as much space as possible when running full screen.  Without it the window size is restricted to multiples of half-size.

Added SBT to file filter, since it can be treated as a disk image.

Eject is now only enabled for drive 2 when a floppy drive is selected, and not for the Atom.

Window messages are now passed to Input::FilterMessage() for all emulator keyboard and mouse processing (UI key processing remains in this module).

Moved frame sync timer/event/callback into OSD.cpp.

Fatal errors no longer call abort() in this module, and are handled more gracefully with exit() in Util.cpp (after flushing modified disk data).

Removed double-sensitivity mouse option, as it was dumb!

Added a bunch of casts to the Win32 message sending used by the UI - Visual Studio .NET is a lot stricter about type sizes, with 64-bit support around the corner.

Changed C++ namespace to class with static functions.