Fix 24-bit mode, which had red and blue the wrong way around.  A previous assumption about surface format in an older Video.cpp masked the problem.
8-bit and 24-bit drawing modes were incorrectly clearing the dirty flags for each line, which prevented the dirty block detection further down from seeing the changed area.  This gave a blank display when running on those modes - it's a bit scary how this has been wrong for so long!

Changed so the SDL cursor is only visible in windowed mode when the mouse isn't being used for emulation and the GUI isn't active.

The back buffer is now cleared whenever the full screen is invalidated, to ensure no garbage is left in the unpainted surface lines when scanlines are enabled.

Added macros for use by 8-bit mode, which previous performed an extra (incorrect) palette look up.