The SDL cursor is now hidden by default, and only displayed when needed (inside Input.cpp).
Input is now updated here when the GUI is active, since the normall processor loop isn't running.  This also removed the need to purge the input queue after unpausing, since key events won't stack up.

The display is now invalidated when we're told part of the view has been exposed and needs painting.