We now pass zero for the display depth when switching to windowed mode, as using the value from SDL_GetVideoInfo() was returning the fullscreen depth when returning from fullscreen mode.  This ensures the new palette is set up correctly for the new depth.
In 8-bit mode we now set the palette on both the front and back surfaces, to ensure fullscreen mode is displayed with the correct palette.

Removed the use of palette offset, as it doesn't seem to be necessary even on Win32 - this also reduces the work done by Display.cpp.

Changed the fullscreen failure fallback to try lower depths before using lower resolutions.