ClipCursor() is now used to prevent the cursor escaping the client area, even when moved very quickly.
Moved the mapped PC keys to a separate structure, removing the overloaded negative values in the structure and matching the SDL version.

The visible state of the cursor is now tracked more carefully, and updated when movement is detected to ensure it's only visible at the appropriate time.  SetCursor keeps a count of the number of calls made so we can't rely on being able to show/hide it at any time, and use a flag to track it ourselves.