Added Home/End/PgUp/PgDn and mouse wheel support for the GUI.
The keyboard modifier state is now reset as part of a keyboard purge, to prevent modifiers getting stuck when a corresponding key-up is not seen.  This can happen when a hot-key or window manager key is used to switch applications.

Symbols generated from the numeric keypad keys are now ignored.