New GUI is now used for file selection when inserting a new disk.  The GUI is also used for various message boxes (mainly not-implemented-yet features!).
actChangeMouse now just toggles between enabled and disabled, as the double-x sensitivity feature has been removed.

The mouse is now automatically grabbed for emulation use when running in fullscreen, as it's not needed for anything else (except the GUI).  The mouse is also released when returning to windowed mode.

Added Scroll Lock as a pause alternatives, since Pause may not come through correctly on all platforms.

Removed InsertDisk() and ResizeWindow() functions as they're no longer needed.