Sound is now muted when the GUI is active.
CTextControl and CCheckBox can now be given an opaque background colour to use.

CEditControl no longer allocates a fixed buffer size since the base SetText() always reallocates it to the required size.  This avoids writing beyond the end of the allocated buffer when adding new characters.

Edit controls now eat cursor movements, to prepare for it processing them to move the cursor.

CMenu now treats accepts a double-click, which is treated as a single click. This allows double-clicks on combo boxes to work correctly and close the on the second click.

CComboBox now has the correct control type, as it was previously a button (copy'n'pasting strikes again!).

CListView now allows a vertical offset of the items, which allow icons with only a single line of text to be centralised in the box.

CListView now fills the background of the selected item in dark blue, to make it stand out more.

Moved dialog centring code to a separate function.  Also revamped the control activation code to ensure one is always active when a sub-dialog is closed.

The cursor keys now move between neighbouring controls of the same type.  This should have been in already for the left/right arrows, but didn't work!

Added override for background erasing, in case the derived control wants to draw it itself.

Removed unused CFileView::DrawItem().