Updated version number on About box and in properties.  Added Dave Hooper back on to credits, since DLL is implicitly linked to released binary.
Added fast-disk-access and sensitivity controls to already squashed Drives options dialog.

Changed manifest to use an ID value instead of the CREATEPROCESS_MANIFEST_RESOURCE_ID value, as the conditionally defining the constant was difficult.  It's available in newer SDKs so it couldn't just be defined as a regular value.  *sighs*