READ_ADDRESS now correctly sets the sector register to hold the track number found in the ID field, after an ID has been read.
Corrected READ_ADDRESS to set BUSY during command execution, as is the case on the real controller.  A bug in SAM DICE relies on it not being set, and it appears to mean it wait for a timeout before reading (?).  I've added a timed workaround to reset busy after 16 status register reads, to keep SAM DICE happy until it's properly implemented.

Changed the floppy pulse signal to be timer based, so it's triggered once per real disk revolution, giving 300 pulses per second.  The actual length of the pulse is a guess for now, and needs official measuring.