SDF images are no longer restricted to 80 tracks, and any whole number of tracks will be accepted.  This is needed to support copy-protected disks that store data on tracks 80-82 (including Lemmings).
When reading a sector, the disk side is no longer compared with the side value in the ID field, as the real controller only checks the track and sector numbers (plus CRC).

The disk spin position now starts at the first sector.

Changed delete to delete[] for some array deletes, to keep valgrind happy.