Controls no longer auto-activate when the mouse moves over them, for a more normal/natural feel.  Clicking on the control activates it, as does tabbing to it using the keyboard.
Added CImageButton control, for a button with an image on it (duh!).

Fixed dialog caption hit test, which was also including the region just below the caption.

Changed return for CDialog::OnMessage, to allow for modal dialog boxes launched from modeless dialogs.  MessageBoxes are now always modal.

The double-click time is reset when the GUI is started, so multiple quick invocations of the debugger (when single stepping) doesn't consider rapid button clicks as a double-click.