Added GetValue() and SetValue() to CWindow, to allow the window text to be manipulated as a UINT value.  CEditControl can also be constructed with a UINT instead of a string, to simplify editable numeric fields.
Added Offset() for a relative window move, SetSize() to resize an existing window, and Inflate() for a relative window resize.

Added support for dialog activation, which is used to dim the fill colour of non-active windows.  This should help the active dialog stand out from larger parent dialogs.

Dialogs are now centralised over the full screen rather than the parent.  This ensures good window placement in cases where the parent has been moved.

Changed to preseve the parent pointer for deleted objects, so it can be used during C++ destruction.

CButton now passes a 1 parameter to the parent OnNotify() if the button was activated with the return key.

Pressing Return in edit controls and radio buttons now notifies the parent of the submission, to allow default button handling in dialogs.

Left/Right cursors can now be used to navigate between radio buttons, in addition to the existing Up/Down cursors.

Increased the gap around CListView labels to improve readability when many labels are wrapped to multiple-lines.