Changes for Mac OS X:
If OpenGL is in use and the the Mac OpenGL library is not detected in the SDL libraries list, add -lGL to use the normal library.

If the Carbon library is detected in the SDL libraries list, assume we're building on OS 10.1 and use cc instead of gcc/g++.

No longer adds /usr/X11R6/lib to the library path, to avoid a linker warning about the path not existing.