We disable Visual Studio 2005's annoying deprecated function warnings unless the standard __USE_SECURE_LIB__ symbol is defined.
access() now only passes through R_OK and W_OK bits, to avoid an annoying Visual Studio 2005 CRT check.  Are they trying to break portability??

Added a compile-time check for DX5, which is needed for DInput joystick definitions.

We define our own LONG_PTR definition if the SDK doesn't seem to be doing it already.  This is needed for SetWindowLongPtr hack to avoid compiler warnings for both 32-bit and 64-bit.