Added new options for: overlay, hdbootrom, asicdelay, samfkeys.  Also added new paths: floppypath, hddpath, rompath, datapath.  Renamed kpminusreset to keypadreset.
Added support for bare disk command-line parameters, which will be inserted into drives 1+2 if found.  If present, drive 1 is also auto-booted without needing to explicitly supply the autoboot option.

autoboot is now a persistent option, used to control autobooting of disks inserted at the boot screen. Previously the value was turned off when the program was closed.

We now accept the Windows' style '/' in addition to '-' for command-line parameters.