The ASIC startup delay is now optional, since not all SAMs seem to suffer from it (reported by Colin Piggot).
The external memory ports are now write-only (thanks to Jiri Veleba).

The auto-boot key tapping is now done in UpdateInput rather than the input port reads, which seems a better place for it.  Auto-booting is now only triggered when we know we're at the startup screen, and is optionally triggered when inserting disks at the boot screen.