The autoboot option/status is now checked when inserting disks into drive 1.
When formatting, we now keep track of the data in the data fields so it can be correctly written to the target disk.  This is needed for the 0xE5 padded Pro-Dos disks to be correctly formatted (as well as the credits and parameters in the first sector).  Thanks to Steve Parry-Thomas for spotting this issue.

Fixed a ~100 byte memory leak for each track formatted.