Added much improved path handling, to give a better starting point for selecting disk images.  If an image is in the drive its existing location is used, otherwise the default disk path is used (can be set in the options).  Disks, hard disks and data files can all have different default paths.
Revamped the import, export, new disk, and HDD properties dialogs for much simpler use.

Fixed MIDI in/out device selection, which was accidentally selecting the 2nd item in the list when no existing value was available.

Removed printer online and print job flushing as part of printing changes, as they should no longer be needed.  At most, the online setting will return in the future if needed.

Fixed minor problems in function key additions when the mapping already existed for a different action.

Decreased mouse hide time to 2 seconds.

Removed e-acute from caption, to match the other versions (where getting the right portable encodings was tricky!).

Updated the disk file filters to include MGT, IMG and CPM.

Added recent disk support to the File menu.

Added menu handling for various options that have been moved out of the regular options dialog.

Added owner-drawn menu support, with a small selection of images in common actions.

The menu is now visible as normal in fullscreen mode, but disappears after the normal mouse hide time or if any key is pressed.

Menu activation is now suppressed if AltGr is currently pressed, to avoid problems using AltGr as the edit key on some Win9x systems.

Alt-Enter is now returned as "processed", to avoid the standard error beep on unknown system key (Alt+???) combinations when toggling fullscreen mode.

We no longer dim the main window when simply inactive (only when a child dialog is active).

To handle window sizes that aren't scale multiples we use a scale value of 0.  This is mapped back to a normal 200% view size if we come to resize the window using the current option setting.  We should probably preserve the window size and position the best we can, though it will be difficult to work in non-Win32 versions.

Removed action support for cases matching the defaults in Action.cpp.

Centred windows are now slighly above centre, which feels more natural.

We now block use of the Windows key when running fullscreen, to avoid accidental presses returning to the desktop.

The Windows key can now be used to toggle the use of the PC function keys as the SAM function keys (Win+F1 = SAM F1).  If the option is already set to use the PC keypad as SAM keys, using the Windows key toggles it back to normal emulation functionality. (Win+F1 = insert floppy 1).

Use of the keypad minus key as reset is now optional, as some users find it dangerously close to F9 for boot (personally I like it for that reason!).

A new GetDlgItemValue() allows any edit control to hold a numeric expression instead of a simple value, and evaluates it when it's fetched.  This is mainly for the Import/Export where calculations can be needed, but can be used elsewhere too.

ShowMessage() now parents over the active window instead of assuming g_hwnd.

We now (optionally) ask the user to confirm saving changes to modified disks.