Drive A: access is now enabled on W2K or above when SAMDISK.SYS is present in the EXE directory.  This is a temporary measure until the new floppy support is added.
We now add the beta date to the About box automatically, using the pre-processor __DATE__ string.

Changed the hard disk dialog to display OK when the filename edit box is empty, as it's a little less worrying that seeing Create by default.

The 8-bit fullscreen Display option is now disabled when the overlay is enabled, since the fullscreen overlay still uses 16-bit mode.

Changed to refresh the SDIDE and YATBus drives when they're changed, to ensure the new option is set correctly.

Failing to save changes now uses the standard Message() mechanism to complain.