Fixed double-increment bug in import/export, which broke sizes over 16K and could cause emulator crashes.
The Insert Disk dialog no longer attempts to use the floppy drive as the current image directory, if the real drive is in use.

Alt-PrintScrn now warns if the overlay is in use when a screenshot is taken.  This may be extended to the non-Alt version if the standard screenshot key is remapped.

We now only grab the mouse input if the GUI isn't active, to avoid a sudden rewarping when the user begins to grab window captions.

Fixed SetTimer/KillTimer mis-use of timer IDs.