Removed SDL OpenGL viewport offset hack
When texture clamping was enabled, the top line of the GUI would blend with the bottom line of the emulated display, giving a visible line.  The hack was to offset by a pixel to hide the line off the bottom of the display.  With the move to a single texture and the removal of clamping we no longer need it.