Improved mouse capture handling
Previously, if the mouse interface was enabled and the window was clicked, mouse control would be acquired.
This happened even if nothing was using the mouse, making it less convenient to leave enabled.

Now if the window is clicked and the mouse hasn't recently been used (except by the ROM), it's not captured.
For the rare use of mouse-controlled BASIC programs, the window can still be double-clicked to force capture.

The mouse interface is now enabled by default, avoiding the inconvenience of rebooting to enable if required.