Symbols are now recognised before literal hex values
Register names that are also hex values are now evaluated as the register.
Adding a leading zero is the easiest way to force it to be seen as a value.

The dot prefix no longer works for either symbols or decimal values.  The dot
suffix still works but only for decimal values.

Also added missing NULL terminator to register array!  The symbol evaluation
order meant this didn't actually cause any problems in earlier versions.