Changed the SBT virtual code file to auto-execute
To aid SAM development, we allow code files to auto-execute on the back
of SimCoupe's support for booting non-bootable disks.  The filename in
the virtual image is now "autoExec", and the execute address of the code
file is set to 32768, to match the load address.

Traditional SBT images have a BOOT signature so the ROM executes them
directly, so this change doesn't affect backwards compatibility.