Removed module initialisation chaining
UI::Init() was called from different locations in Win32 and SDL, which
didn't feel right. It's now called from Main::Init(), like other modules.

OSD::Init() is also performed first, as it used to be. This gives the
OS-specific module first chance to perform any special initialisation.

This also fixes a bug with portable mode under Win32, where Frame was
initialised with default options, and other modules had the correct
options. If the border area was not the default value, the main window
would be created the wrong size for the display area (thanks Chris!).