Added initial support for shared resource files
Resource files are read-only data files supplied with SimCoupe, to
be shared by all users. This currently covers the symbol maps and
SDL icon, but may also be used for ROM images in the future. MFP_EXE
has been renamed to MFP_RESOURCE to support them.

Under Windows resources will be stored with SimCoupe.exe, which
may be in a read-only location such as Program Files.

Under Mac OS X we use SDL_GetBasePath(), which will return the
path of the Resources directory in the app bundle. This requires
SDL 2.0.1, but that should always be the case for Mac builds.

Under Linux we use ${CMAKE_INSTALL_PREFIX}/share/simcoupe/ instead.
SDL_GetBasePath() may work, but its return doesn't seem so well
defined.