Removed support for low-resolution rendering
This was originally added as as a drawing optimisation, because modes other
than 3 require each pixel to be doubled horizontally. The core drawing hid
this using a template class, but anything else dealing with pixels had to
worry about the difference.

The optimistion was needed for my Pocket PC device back in 2002, with its
low-resolution display and slow CPU. Modern mobile devices can display full
resolution (and then some), and shouldn't notice the extra processing.