Fixed invalid CHS mapping with large disks (8GB+)
Disks beyond 16383*16*63 sectors should always be reported with a
CHS geometry of 16383,16,63, according to the ATA specs. The previous
attempt to extend the geometry was wrong. Accessing beyond 8GB requires
LBA instead of CHS.

This change restores the 16383 limit on cylinder count, and removed
heads=255 option. This also brings it into line with recent changes in
SAMdisk geometry calculation.

There's still a small risk the invented geometry will not match the
true geometry, causing BDOS to not recognise the disk. This is made
worse by the AL BDOS using a slightly different calculation. There's
no solution, as unlike BDOS, we can't see the CHS geometry in the
IDENTIFY area of CF cards in a USB card reader.