Reduced sound latency to 3 frames
5 frames felt excessive, but additional buffers seem to be needed
under newer versions of Windows, due to a slow release of played
buffers that delayed the write cursor being updated. I don't
remember this being a problem under Windows XP, and it may be due
to DirectSound effectively being emulated now. We may need to use
XAudio2 to avoid it.