Fixed Atom media detection and disabled byte-swapping
Atom media detection now correctly uses the CHS geometry to
calculate the position of the first sector, rather than the total
LBA sector count.

We now require media to match the appropriate Atom interface type,
instead of automagically adapting to it. Mixing disk types was
intended to be a convenience, but it was more likely to confuse
users -- especially those sharing media with real SAM hardware.

The autobyteswap option has also been removed.