Fixed build errors with gcc 7.3 under Ubuntu 18.04
Renamed CHAR_WIDTH to CHR_WIDTH, to avoid clash with new macro
defined in limits.h as part of ISO/IEC TS 18661-1:20.

Increased format buffer width to avoid losing the last character
of the day in the month when we reach the year 10,000. I guess I
did ask for the compiler to be pedantic.