Relax resource location under Windows
RESOURCE_DIR is defined to be the resource installation location
at build time. This works fine for non-Windows platforms, where
the locations are generally fixed, but doesn't work if SimCoupe
is on a USB stick (portable mode). RESOUCE_DIR still works fine
for Windows development, after an initial 'install' action.

If the resource file doesn't exist we'll look in the directory
containing the EXE, regardless of what the CWD is.