Removed Allegro target
The Allegro code was based on the old 4.x API, which is not
compatible with the current 5.x API. We really only added this
for the DOS version, which is not really of interest anymore.

Should the Allegro version be needed again it'll be built up
using the old code as a reference. At this time it's too much
trouble to try and keep in sync. Once we have more generic
back-ends for the different sub-systems it may be more tempting.