Fixed filesystem include check
We need to use the C++ header check to find it, and the angled
brackets should not be given. C++17 is also set globally before
the check as it may be required in some environments.

Moved NOMINMAX macro to SimCoupe.h to ensure it's set before
Windows headers can be included. Otherwise std:min and std::max
generate build errors.