Fixed passing bare disk paths to portable version
The first non-option argument passed on the command-line should be
inserted into drive 1 and auto-booted. This worked for the non-portable
version, where a SimCoupe.cfg wasn't found in the EXE folder (Win32).

With portable installations the config file was reloaded and the
command-line re-parsed, but due to a static variable the drive 1 slot
was considered filled so the disk was inserted into drive 2 instead.
We now re-initialise options and set back to drive 1 on each re-load.