Added support for saving SSX screen files
These files hold the display memory data for the main screen area,
followed by the CLUT indices into the 128 SAM palette colours. They are
static image formats intended for preservation use.

Mid-display changes to VMPR or CLUT will give the wrong result for the
formats above, and if detected a raster-accurate representation of the
display is saved instead (512x192 palette indices).

A new SaveSSX action (51) is available for key bindings. Two screenshot
saving items are available on the Record menu in the Windows version.