Fixed line interrupt issue on reset
The line interrupt port isn't cleared by a reset, but CPU::Reset was
removing the event for the next trigger point. If the next line number
written after reset was the same as the previous value the handler
ignored it, preventing a new line interrupt event being set up.

If line 0 was active at the point of the reset, the ROMs setting line 0
had no effect, resulting in the boot screen stripes not working.
Example: POKE 32768,243,175,211,249,118 : CALL 32768  then reset.