Fixed crash opening empty tape image (#34)
libspectrum allowed the empty file to be opened, but any attempt to
read the current file position resulted in a call to abort(). We're
using quite an old version so this may have already been fixed.

Even avoiding that, an empty image would mean no entries in the tape
browser. To avoid confusion we now reject zero-sized images.