Use C-style file size due to path encoding issues
ghc::filesystem::file_size expects the path to be UTF-8, but paths under
Windows that came from the Win32 API will be using the default codepage,
which is not UTF-8. Paths containing non-ASCII characters will not be
recognised correctly. In this case the file size of such paths will be seen
as zero. To work around this we'll use the old CRT method for now.

We're still using a mix of old and new file APIs as the codebase is slowly
modernised. There are likely to be other lurking path issues, mostly under
Windows. Ideally the core will use UTF-8 and we'll convert at any API
boundaries that use other encodings. This will likely affect file APIs and
user interfaces.