Fix VoiceBox detecting when disabled (#40)
Setting voicebox=0 only disabled its sound mixing so it couldn't be heard,
but it was still responding to all other I/O. This meant it would still be
detected by the appropriate port reads.

We still process port writes to leave the device in a better state should it
be re-enabled at runtime, but no longer respond to port reads.