Fixed frozen screen after keyboard speedup (#35)
Increasing emulation speed using the keyboard caused all frame drawing to be
disabled, making the emulator appear frozen. To recover it required a reset
of the emulated SAM.

Speed changes were clearing any active CPU reset condition. If the fast
reset option was enabled this disabled all frame drawing, to speed through
the slow ROM memory test. Drawing was only re-enabled when the keyboard was
next read.

The flag to skip frame drawing was being set on any clear of the reset
signal. If the running code was not polling the keyboard it would be stuck
in no-draw mode. We now set the flag only when changing from reset asserted
to reset clear, which will be the start of a genuine reset.