Hide OS key combinations
If our application still has the focus when a Windows key combination is
used we may process one or both key presses in the emulation. We now skip
input processing if the Windows key is pressed, unless it's used in
combination with a function key to access SAM function keys.

We also ignore Alt-Tab, which should be left for exclusive OS use.