Changed to pull SAASound from GitHub
- replaced Extern\SAASound with FetchContent from GitHub (#30)
- bump required CMake to 3.14 for FetchContent_MakeAvailable
- added find_package module wrapper for SAASound
- changed to use module versions of find_package libraries
- marked all library imports as PRIVATE

FetchContent requires Internet access at CMake *configure* time to fetch the
SAASound project, but it'll be cached and reused for later builds.
GIT_REPOSITORY can be pointed at a local path and/or GIT_TAG at a branch
name for experimental builds. We're pulling the current HEAD commit from the
master branch, but this can be updated to select a later version as needed.

Other external libraries will probably be moved to use the same mechanism.
It will avoid some cases of having to disable emulator features due to
missing libraries since they can be pulled in if not available locally.

Also changed to be less forceful with installation folders. Win32 debugging
still requires at least one "Install simcoupe" from the Build menu to copy
the resources, but the normal version will look in the EXE directory if it
can't find them in the build-time resource location.