Changes to border area options
- removed "Normal", which didn't really offer anything unique.
- increased TV Visible vertical lines to cover the recommended Action-safe
area, which is 93% of the full width/height. Most TVs should cover this.
- replaced "Complete" with "Full Active", to show the full active region
from the PAL display output. Anything seen in this mode could be seen on
a good TV or monitor, so developers should avoid flicker here!

Also changed 'tvaspect' option to be enabled by default. Combining this with
TV Visible borders gives a good representation of the SAM display output.