Improved frequency of generated sound output (#36)
The internal sound clock was was ~44128 Hz, which didn't match the 44100 Hz
selected for the output device. This mismatch was inaudible for normal
playback but could be detected if the SAA was driven for sample playback.
Special thanks to AntonyDTA for reporting and helping reproduce this.

Blip_Buffer has been updated to support 64-bit types, thanks to work done by
the Mednafen team. This allows us to increase the accuracy of its internal
frequency ratios, improving the output frequency to 44100.0005 Hz. This
should hopefully avoid most phasing issues.

Changed to use the new SAASound function to set frequency, and removed
frequency as a limiting factor on some operations. Also removed mentions of
avireduce sound option that was removed a while ago.