Improved bus values from unhandled ports
Rather than always returning 0xff we return 0xff in the border and the ATTR
value over the main screen. This is closer to what the real system does but
not yet perfect. The real system values appear to be deterministic.

This change means KEDisk now starts correctly. It was unintentionally using
port 0 instead of the real FDC base port for the first reads. Our 0xff value
was seen as the FDC being permanently busy and it waited forever. The mix of
values allows it to continue and set up the port for later reads.

Port reads now return early, with only unhandled reads hitting the bottom of
IO::In. Debugger tracking of port I/O was also moved to the CPU module.