Added samdos2 symbols to debugger (#44)
When paged into section B, symbols for the standard SAMDOS v2.0 image are
now shown. Basic version and signature checks are made to allow versions
with minor patches to be recognised.

Also fixed name-to-value mapping being cleared when updating symbols. This
caused symbol lookups to work only the first time the debugger was opened!