Fixed EDSK save attempt corrupting file (#57)
EDSK saving relied on stream rewind succeeding, but it failed due to the
underlying stream being closed. The stream is closed when the drive motor
switches off to aid image sharing with other applications. Reading and
writing will reopen the file at offset 0 so the rewind shouldn't be failed.

Changed Rewind to not report a status, and EDSK no longer attempts to check
it, allowing the save to continue like the other disk image types.