GIF recording enhancements (#59)
- added option to record with full frame size
- added framerate control, with choice of 50(default)/25/16.7/12.5 fps

We were previously recording at half-size and limiting to 16.6fps so it
played in IE6. Modern versions of Firefox/Chrome/Edge all seem to play
full-size 50fps animations without a problem, so this is the default.

Refactored some of the GIF code to modernise it a little. Plenty left to
improve in the future though/