Fix read-only error writing with Pro-Dos v2 (#64)
The issue was caused by a timeout condition falsely triggering, which
happens when DRQ is set but data isn't being read. SAM DICE uses this to
synchronise with the track index.

In this case DRQ was set for data to be written, and the lack of data
_reads_ caused SimCoupe to generate a LOST DATA error. The timeout counter
also needed to be reset when data is _written_ to avoid the condition.

Pro-Dos does a better job of checking the flags after commands than SAMDOS,
which is perhaps why this hadn't been seen until now.