Fix command-line disk auto-booting (#69)
The auto-boot was correctly queued, but the READKEY loop at the copyright
message doesn't read LAST_K input. It was queued when we weren't at the
copyright message so we're not able to skip the loop.

The process has been simplified to only attempt the boot when READKEY is
called from the expected place in ROM 0. This also clears the boot attempt
so it won't turn up later, or be attempted when not at the boot screen.