Expose CPU halted status as variable
The 'halted' variable returns 1 when the CPU is halted, 0 otherwise. This
can be useful for some debugger expressions.

Also changed breakpoints to only match PC when the CPU isn't halted. This
ensures execution breakpoints stopping when an instruction is really about
to be executed, rather than when first halted. This doesn't make a
difference with the current CPU code but will once HALT behaviour is fixed.