Updated to use vcpkg version of xaudio2redist
We now look for the vcpkg version of xaudio2redist, which is needed to
support Windows 7. The fallback to use the DirectX SDK headers is no longer
supported. If not building for Windows 7 we will still fall back on the
Windows SDK version of XAudio2.

Fixed a calling convention issue when using the Platform SDK headers with
the 32-bit Windows build. Using decltype to extract the prototype from
XAudio2Create was missing the required __stdcall attribute.