Added frame limit to fast reset
We used to wait for the first keyboard read to re-enable the display after
a fast reset, if it was enabled. If the ROM was invalid/corrupt this never
happened, leading to the display never being redrawn.

Display updates are re-enabled 200 frames after a reset, or when the ROM
copyright message is drawn, whichever happens first.

Also limited work skipped during fast reset to final frame presentation.